Rutas y MainScreens refactorizadas
